fix-fe: review corrections
