+++ apps/web/src/data/profile.ts
@@ -1,0 +1,10 @@
+import { NomineeProfile } from "@/types/profile.type";
+
+export const PROFILE_ATTRIBUTES: {
+  label: string;
+  key: keyof Pick<NomineeProfile, "field" | "region" | "organization">;
+}[] = [
+  { label: "Field", key: "field" },
+  { label: "Region", key: "region" },
+  { label: "Organization", key: "organization" },
+];
