+++ apps/web/src/types/profile.type.ts
@@ -25,23 +25,3 @@ export interface ProfileLinksProps {
   title: string;
   links: string[];
 }
-
-export interface ApiNominee {
-  id: number;
-  first_name: string;
-  last_name: string;
-  field: string;
-  country: string;
-  organization?: string;
-  profile_image_url: string;
-}
-
-export interface ApiNomination {
-  id: number;
-  nominee_id: number;
-  status: string;
-  description: string;
-  evidence_urls: string | string[];
-  supporting_urls: string | string[];
-  nominee: ApiNominee;
-}
